Add symlink to al5poly DLL into bin during make.
This way you can successfully `make clean run` without having to
manually copy or symlink.

It should be an MSYS symlink, which probably would only work from a pure
MSYS environment, but for some reason it appears to be creating a copy
instead. Which is OK. It works in a normal Windows environment too. At
the expense of a small bit of duplication (which is nothing new for
Windows).