h4x: Add libal5poly dependency to Makefile.
I am currently in Linux so I can't fully test this to completion, but
this should at least be the beginning of automagically pulling in the
dependency (given a Git installation) and building it (given a MinGW
installation). I'll need platform specifics to make it work in both
Windows and Linux, and I don't have the experience or time to do that
right now. :P