Add build-deps target to Makefile.
This target will always build the dependencies, regardless of whether
their outputs exist. It might be best to make that the default, but I
think that MinGW/MSYS is having trouble identifying when files change
and seems to always rebuild the whole project. Building libal5poly each
time in addition to the game would slow me down that much more.