Update input processing to handle repeats.
In order to handle repeats the input manager has to track key states
itself and send events separately. So we end up with dummy keyPress and
keyRelease that merely track state, and then sendEvents that actually
sends events to the event handlers for processing. It's not yet a
perfect system. Need a way to track which events are repeats and which
are not so that event handlers can either process them or ignore them...
Though somehow magically the jumping logic seems to already work this
way. I'm going to say it is accidental...